// src/api/models/articleResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FlowNews API
 * OpenAPI spec version: v1
 */

/**
 * 기사 정보
 */
export interface ArticleResponse {
  /** 기사 ID */
  id: number;
  /** 기사 제목 */
  title: string;
  /** 기사 출처(언론사) */
  source: string;
  /** 기사 링크 */
  url: string;
}